+++ CHANGELOG.md
@@ -6,12 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.6] - 2022-MM-DD aka "*Forced* to eternal maintenance."
+## [0.2.6] - 2022-MM-DD aka "Things you'll never notice."
 ☮️ Peace in the world, or the world in pieces. 🕊️
 ### Added
 - Cold Bodied to the default mapping (and thus chat output) and an icon that is in line with the rest of SUCCs icons.
 ### Changed
 - SUCC now uses the `actor.toggleConviction()` function of SWADE v. 2.1.0 instead of recreating it.
+- SUCC now lets the system figure out the proper values for boost/lower in case of going above a d12 / below a d4.
 
 ## [0.2.5] - 2022-11-07 aka "Exporting bugs since 2022."
 ### Fixed